+++ Sistema.Web/Models/Almacen/Almacen/ActualizarViewModel.cs
@@ -1,3 +1,4 @@
+using System;
 using System.ComponentModel.DataAnnotations;
 
 namespace Sistema.Web.Models.Almacen.Almacen
@@ -8,6 +9,6 @@ namespace Sistema.Web.Models.Almacen.Almacen
         public int AlmId { get; set; }
         public string AlmDescripcion { get; set; }
         public char AlmEstatus { get; set; }
-        public int SucId { get; set; }
+        public Int16 SucId { get; set; }
     }
 }
