+++ Sistema.Web/Models/Almacen/Almacen/SelectViewModel.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace Sistema.Web.Models.Almacen.Almacen
 {
     public class SelectViewModel
@@ -5,7 +7,7 @@ namespace Sistema.Web.Models.Almacen.Almacen
         public int AlmId { get; set; }
         public string AlmDescripcion { get; set; }
         public char AlmEstatus { get; set; }
-        public int SucId { get; set; }
+        public Int16 SucId { get; set; }
 
         public string sucursal { get; set; }
     }
