+++ Sistema.Web/Models/Almacen/Sucursal/SucursalViewModel.cs
@@ -1,8 +1,10 @@
+using System;
+
 namespace Sistema.Web.Models.Almacen.Sucursal
 {
     public class SucursalViewModel
     {
-        public int SucId { get; set; }
+        public Int16 SucId { get; set; }
         public string SucNombre { get; set; }
     }
 }
